Rename classes and add test packages
